📦 NEW: add pages structure
