+++ jest.config.js
@@ -5,7 +5,6 @@ module.exports = {
 		'src/filter.json',
 		'vite-env.d.ts',
 		'cypress',
-		'test-util.tsx',
 	],
 	coverageReporters: ['json', 'text', 'lcov', 'clover'],
 	coverageThreshold: {
