+++ package.json
@@ -79,8 +79,8 @@
 		"stylelint-config-recommended": "^5.0.0",
 		"stylelint-config-styled-components": "^0.1.1",
 		"stylelint-processor-styled-components": "^1.10.0",
-		"typescript": "^3.9.10",
-		"vite": "2.3.8"
+		"typescript": "^4.4",
+		"vite": "2.5.6"
 	},
 	"prettier": {
 		"singleQuote": true,
