+++ tsconfig.json
@@ -11,9 +11,9 @@
 		"module": "CommonJS",
 		"moduleResolution": "Node",
 		"resolveJsonModule": true,
-		"isolatedModules": true,
+		"isolatedModules": false,
 		"noEmit": true,
-		"jsx": "react",
+		"jsx": "react-jsx",
 		"types": [
 			"jest",
 			"react",
